[wip] additional lowering functionality
