+++ tests/test_unittest/test_gtc/test_oir_to_npir.py
@@ -14,6 +14,12 @@ EMPTY_VERTICAL_LOOP = oir.VerticalLoop(
 )
 
 
+EMPTY_HORIZONTAL_EXECUTION = oir.HorizontalExecution(
+    body=[],
+    mask=None,
+)
+
+
 def test_stencil_to_computation():
     stencil = oir.Stencil(
         name="stencil",
@@ -34,3 +40,18 @@ def test_stencil_to_computation():
     assert computation.field_params == ["a"]
     assert computation.params == ["a", "b"]
     assert len(computation.vertical_passes) == 1
+
+
+def test_vertical_loop_to_vertical_pass():
+    vertical_loop = oir.VerticalLoop(
+        interval=oir.Interval(
+            start=oir.AxisBound.start(),
+            end=oir.AxisBound.end(),
+        ),
+        horizontal_executions=[],
+        loop_order=common.LoopOrder.PARALLEL,
+        declarations=[],
+    )
+    vertical_pass = OirToNpir().visit(vertical_loop)
+
+    assert vertical_pass.body == []
